refactor: move algorithm to a crate
