+++ .taplo.toml
@@ -2,9 +2,16 @@
 indent_string = "    "
 
 [[rule]]
-keys = ["dependencies", "*-denpendencies", "lints", "patch.*", "profile.*"]
-
-[rule.formatting]
-reorder_keys = true
-reorder_arrays = true
-align_comments = true
+include = ["**/Cargo.toml"]
+keys = [
+    "dependencies",
+    "dev-dependencies",
+    "build-dependencies",
+    "target.*.dependencies",
+    "lints",
+    "patch.*",
+    "profile.*",
+    "workspace.dependencies",
+    "lints.dependencies",
+]
+formatting = { reorder_keys = true, reorder_arrays = true, align_comments = true }
