+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "vchord"
 version.workspace = true
-edition.workspace = true
+edition = "2021"
 
 [lib]
 name = "vchord"
@@ -20,25 +20,21 @@ pg16 = ["pgrx/pg16", "pgrx-catalog/pg16"]
 pg17 = ["pgrx/pg17", "pgrx-catalog/pg17"]
 
 [dependencies]
-always_equal = { path = "./crates/always_equal" }
+algorithm = { path = "./crates/algorithm" }
 distance = { path = "./crates/distance" }
-rabitq = { path = "./crates/rabitq" }
+k_means = { path = "./crates/k_means" }
 random_orthogonal_matrix = { path = "./crates/random_orthogonal_matrix" }
 simd = { path = "./crates/simd" }
 vector = { path = "./crates/vector" }
 
 half.workspace = true
-log = "0.4.25"
 paste = "1"
 pgrx = { version = "=0.12.9", default-features = false, features = ["cshim"] }
 pgrx-catalog = "0.1.0"
 rand.workspace = true
-rayon = "1.10.0"
 serde.workspace = true
 toml = "0.8.19"
-validator = { version = "0.19.0", features = ["derive"] }
-zerocopy = "0.8.14"
-zerocopy-derive = "0.8.14"
+validator.workspace = true
 
 [patch.crates-io]
 half = { git = "https://github.com/tensorchord/half-rs.git", rev = "3f9a8843d6722bd1833de2289347640ad8770146" }
@@ -56,14 +52,19 @@ edition = "2021"
 
 [workspace.dependencies]
 half = { version = "2.4.1", features = ["serde", "zerocopy"] }
+log = "0.4.25"
 rand = "0.8.5"
 serde = "1"
+validator = { version = "0.20.0", features = ["derive"] }
+zerocopy = "0.8.14"
+zerocopy-derive = "0.8.14"
 
 [workspace.lints]
 clippy.identity_op = "allow"
 clippy.int_plus_one = "allow"
 clippy.needless_range_loop = "allow"
 clippy.nonminimal_bool = "allow"
+rust.unsafe_code = "deny"
 rust.unsafe_op_in_unsafe_fn = "deny"
 rust.unused_lifetimes = "warn"
 rust.unused_qualifications = "warn"
