+++ crates/algorithm/src/freepages.rs
@@ -1,9 +1,9 @@
-use crate::algorithm::tuples::*;
-use crate::algorithm::*;
-use crate::utils::pipe::Pipe;
+use crate::pipe::Pipe;
+use crate::tuples::*;
+use crate::*;
 use std::cmp::Reverse;
 
-pub fn mark(relation: impl RelationWrite, freepage_first: u32, pages: &[u32]) {
+pub fn mark(index: impl RelationWrite, freepage_first: u32, pages: &[u32]) {
     let mut pages = pages.to_vec();
     pages.sort_by_key(|x| Reverse(*x));
     pages.dedup();
@@ -18,7 +18,7 @@ pub fn mark(relation: impl RelationWrite, freepage_first: u32, pages: &[u32]) {
             }
             local
         };
-        let mut freespace_guard = relation.write(current, false);
+        let mut freespace_guard = index.write(current, false);
         if freespace_guard.len() == 0 {
             freespace_guard.alloc(&serialize(&FreepageTuple {}));
         }
@@ -30,19 +30,19 @@ pub fn mark(relation: impl RelationWrite, freepage_first: u32, pages: &[u32]) {
             freespace_tuple.mark(local as _);
         }
         if freespace_guard.get_opaque().next == u32::MAX {
-            let extend = relation.extend(false);
+            let extend = index.extend(false);
             freespace_guard.get_opaque_mut().next = extend.id();
         }
         (current, offset) = (freespace_guard.get_opaque().next, offset + 32768);
     }
 }
 
-pub fn fetch(relation: impl RelationWrite, freepage_first: u32) -> Option<u32> {
+pub fn fetch(index: impl RelationWrite, freepage_first: u32) -> Option<u32> {
     let first = freepage_first;
     assert!(first != u32::MAX);
     let (mut current, mut offset) = (first, 0_u32);
     loop {
-        let mut freespace_guard = relation.write(current, false);
+        let mut freespace_guard = index.write(current, false);
         if freespace_guard.len() == 0 {
             return None;
         }
