+++ crates/algorithm/src/pipe.rs
