+++ rustfmt.toml
@@ -1,1 +1,2 @@
 style_edition = "2024"
+imports_granularity = "Module"
