+++ src/bin/pgrx_embed.rs
@@ -1,1 +1,2 @@
+#![allow(unsafe_code)]
 ::pgrx::pgrx_embed!();
