+++ src/datatype/memory_vector.rs
@@ -1,19 +1,13 @@
-use pgrx::datum::FromDatum;
-use pgrx::datum::IntoDatum;
-use pgrx::pg_sys::Datum;
-use pgrx::pg_sys::Oid;
-use pgrx::pgrx_sql_entity_graph::metadata::ArgumentError;
-use pgrx::pgrx_sql_entity_graph::metadata::Returns;
-use pgrx::pgrx_sql_entity_graph::metadata::ReturnsError;
-use pgrx::pgrx_sql_entity_graph::metadata::SqlMapping;
-use pgrx::pgrx_sql_entity_graph::metadata::SqlTranslatable;
+use pgrx::datum::{FromDatum, IntoDatum};
+use pgrx::pg_sys::{Datum, Oid};
+use pgrx::pgrx_sql_entity_graph::metadata::*;
 use std::marker::PhantomData;
 use std::ptr::NonNull;
 use vector::VectorBorrowed;
 use vector::vect::VectBorrowed;
 
 #[repr(C, align(8))]
-pub struct VectorHeader {
+struct VectorHeader {
     varlena: u32,
     dims: u16,
     unused: u16,
@@ -27,10 +21,10 @@ impl VectorHeader {
         }
         (size_of::<Self>() + size_of::<f32>() * len).next_multiple_of(8)
     }
-    pub unsafe fn as_borrowed<'a>(this: NonNull<Self>) -> VectBorrowed<'a, f32> {
+    unsafe fn as_borrowed<'a>(this: NonNull<Self>) -> VectBorrowed<'a, f32> {
         unsafe {
             let this = this.as_ptr();
-            VectBorrowed::new_unchecked(std::slice::from_raw_parts(
+            VectBorrowed::new(std::slice::from_raw_parts(
                 (&raw const (*this).elements).cast(),
                 (&raw const (*this).dims).read() as usize,
             ))
@@ -92,7 +86,7 @@ impl VectorOutput {
     pub fn as_borrowed(&self) -> VectBorrowed<'_, f32> {
         unsafe { VectorHeader::as_borrowed(self.0) }
     }
-    pub fn into_raw(self) -> *mut VectorHeader {
+    fn into_raw(self) -> *mut VectorHeader {
         let result = self.0.as_ptr();
         std::mem::forget(self);
         result
