--- src/upgrade/mod.rs
@@ -1,1 +1,0 @@
-mod symbols;
