--- src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod k_means;
-pub mod parallelism;
-pub mod pipe;
